add dark mode and RehydrateStorage in store
