+++ app/globals.css
@@ -1,4 +1,5 @@
 @import "tailwindcss";
+@custom-variant dark (&:where(.dark, .dark *));
 
 :root {
   --font-sans: var(--font-geist-sans);
@@ -7,4 +8,9 @@
 
 body {
   font-family: var(--font-sans), Arial, Helvetica, sans-serif;
+}
+
+@utility container {
+  margin-inline: auto;
+  max-width: 600px;
 }
