+++ app/layout.tsx
@@ -24,11 +24,7 @@ export default function RootLayout({
   children: React.ReactNode;
 }) {
   return (
-    <html
-      lang="en"
-      suppressHydrationWarning
-      className="transition-colors duration-300"
-    >
+    <html lang="en" className="transition-colors duration-300">
       <body
         className={`${geistSans.variable} ${geistMono.variable} antialiased transition-colors`}
       >
