+++ app/not-found.tsx
@@ -1,8 +1,6 @@
 "use client";
 import { Button } from "@mui/material";
 import NextLink from "next/link";
-
-// next
 import Head from "next/head";
 
 export default function Page404() {
@@ -11,14 +9,9 @@ export default function Page404() {
       <Head>
         <title> not found Page </title>
       </Head>
-      <div className="text-center mt-10">
+      <div className="text-center pt-10">
         <h1 className="mb-8"> 404 not found Page</h1>
-        <Button
-          component={NextLink}
-          href="/panel"
-          size="large"
-          variant="contained"
-        >
+        <Button component={NextLink} href="/" size="large" variant="contained">
           home
         </Button>
       </div>
