+++ postcss.config.mjs
@@ -1,11 +1,6 @@
 const config = {
-  plugins: ["@tailwindcss/postcss"],
-};
-
-export default config;
-module.exports = {
   plugins: {
-    tailwindcss: {},
-    autoprefixer: {},
+    "@tailwindcss/postcss": {},
   },
 };
+export default config;
